refactor(utils/getUserPoints) переименовал getUserScore в getUserPoints
